requirements.txt
Django==1.9.7
gunicorn==19.6.0
nose==1.3.7
psycopg2==2.6.1
PyGithub==1.26.0
requests==2.10.0
